+++ src/Redmine/Client/Psr18Client.php
@@ -118,12 +118,17 @@ final class Psr18Client implements Client, HttpClient
     /**
      * Create and send a GET request.
      *
+     * @deprecated v2.8.0 Use `\Redmine\Http\HttpClient::request()` instead
+     * @see \Redmine\Http\HttpClient::request()
+     *
      * @throws ClientException If anything goes wrong on the request
      *
      * @return bool true if status code of the response is not 4xx oder 5xx
      */
     public function requestGet(string $path): bool
     {
+        @trigger_error('`' . __METHOD__ . '()` is deprecated since v2.8.0, use `\Redmine\Client\Psr18Client::request()` instead.', E_USER_DEPRECATED);
+
         $response = $this->runRequest('GET', $path);
 
         return $response->getStatusCode() < 400;
@@ -132,12 +137,17 @@ final class Psr18Client implements Client, HttpClient
     /**
      * Create and send a POST request.
      *
+     * @deprecated v2.8.0 Use `\Redmine\Http\HttpClient::request()` instead
+     * @see \Redmine\Http\HttpClient::request()
+     *
      * @throws ClientException If anything goes wrong on the request
      *
      * @return bool true if status code of the response is not 4xx oder 5xx
      */
     public function requestPost(string $path, string $body): bool
     {
+        @trigger_error('`' . __METHOD__ . '()` is deprecated since v2.8.0, use `\Redmine\Client\Psr18Client::request()` instead.', E_USER_DEPRECATED);
+
         $response = $this->runRequest('POST', $path, $body);
 
         return $response->getStatusCode() < 400;
@@ -146,12 +156,17 @@ final class Psr18Client implements Client, HttpClient
     /**
      * Create and send a PUT request.
      *
+     * @deprecated v2.8.0 Use `\Redmine\Http\HttpClient::request()` instead
+     * @see \Redmine\Http\HttpClient::request()
+     *
      * @throws ClientException If anything goes wrong on the request
      *
      * @return bool true if status code of the response is not 4xx oder 5xx
      */
     public function requestPut(string $path, string $body): bool
     {
+        @trigger_error('`' . __METHOD__ . '()` is deprecated since v2.8.0, use `\Redmine\Client\Psr18Client::request()` instead.', E_USER_DEPRECATED);
+
         $response = $this->runRequest('PUT', $path, $body);
 
         return $response->getStatusCode() < 400;
@@ -160,12 +175,17 @@ final class Psr18Client implements Client, HttpClient
     /**
      * Create and send a DELETE request.
      *
+     * @deprecated v2.8.0 Use `\Redmine\Http\HttpClient::request()` instead
+     * @see \Redmine\Http\HttpClient::request()
+     *
      * @throws ClientException If anything goes wrong on the request
      *
      * @return bool true if status code of the response is not 4xx oder 5xx
      */
     public function requestDelete(string $path): bool
     {
+        @trigger_error('`' . __METHOD__ . '()` is deprecated since v2.8.0, use `\Redmine\Client\Psr18Client::request()` instead.', E_USER_DEPRECATED);
+
         $response = $this->runRequest('DELETE', $path);
 
         return $response->getStatusCode() < 400;